Created a producer wrapper and a utility class.
