+++ src/main/java/dms/testkafkaapp/Constants.java
@@ -1,0 +1,6 @@
+package dms.testkafkaapp;
+
+public class Constants {
+    public static final String TOPIC_NAME = "test";
+    public static final int TIMEOUT_MS = 10000;
+}
